Put the whole bind flow in one modal, off the wall
Binding was scattered across two surfaces: dashed ghost tiles on the deck
took the PIN and reported failures, while a separate BIND sheet explained
what a PIN was, how to install the CLI, and offered scan/paste. One task,
two places, and the wall asking a question the wall couldn't answer.

Add Host is now the single front door. It opens on a BIND | MANUAL choice
bar — BIND first, since the deck's BIND chip is gone and one `+ HOST` chip
is the only entry. Editing a host shows no bar; there is nothing to bind.

The Bind pane reads in the order a first-timer needs: the commands to run
on the machine (Homebrew, curl, `mpx bind` — each with a COPY button, via
the same `HostGuide`/`CopyableCommandField` the tmux-install and keychain
sheets use), the machines answering, scan/paste for a machine this network
can't hear, then the key promise. A candidate is a form row carrying the
machine's own words — user@name, address, SSH fingerprint — so it can be
checked against the terminal before anything is enrolled.

Discovery follows the surface: `BindPane` sets `bindSurfaceOpen`, the deck
still owns begin/end, and browsing runs only while that flag is set or an
enrollment is in flight (the handshake needs the live browser for its
endpoint). In-flight specifically, not merely a non-empty list, or a
candidate parked on FAILED would hold the browser open all session behind
a closed modal. So the local-network prompt can now only ever follow an
explicit ADD HOST tap, on any launch, fleet or no fleet.

What that costs, deliberately: "run one command and glance at the deck" is
no longer true — you glance at the modal. That promise is what the looser
mid-build gating existed to buy, and it goes with the ghost tiles. The
empty wall's hint changes with it, since pointing at `mpx bind` would now
promise a tile that never arrives.

The paywall and the "not a bind code" alert move into the modal too: an
alert raised on the deck behind an open sheet is never seen.

Verified on the iPad simulator against the dev harness, clean install each
time: `multiplex:bind:335c03ee:ipad-pro-13-inch-m5` landed in the harness's
authorized_keys through the real discovery → PIN → handshake path, and the
bound host came up live on the wall. The negative was proven from the log
too — two hosts, deck frontmost, no pane: zero `bind browse started`;
opening the pane: exactly one. visionOS renders the new modal correctly
(its own chassis sheet, no scan affordance — no camera access there).
544 XCTest + 29 swift-testing pass; both platforms build.
