+++ Multiplex/Views/Deck/AddHostSheet.swift
@@ -2,13 +2,27 @@ import SwiftUI
 import UIKit
 
 /// Add or edit a host. Secrets go straight to the Keychain on save.
+///
+/// Adding is two roads to the same place, so they share one modal behind a
+/// choice bar: BIND (run one command on the machine and confirm it) and
+/// MANUAL (type the destination). Editing an existing host has only the form
+/// — there is nothing to bind — so the bar is absent there entirely.
 struct AddHostSheet: View {
     @Environment(HostStore.self) private var store
     @Environment(EntitlementStore.self) private var entitlements
+    @Environment(BindController.self) private var bind
     @Environment(\.dismiss) private var dismiss
 
     var editing: Host?
 
+    /// Which road this modal is showing. Add opens on BIND: the deck no
+    /// longer carries a BIND chip, so this bar is where the flow lives.
+    enum Mode: Hashable {
+        case bind
+        case manual
+    }
+
+    @State private var mode: Mode = .bind
     @State private var name = ""
     @State private var hostname = ""
     @State private var port = "22"
@@ -84,16 +98,18 @@ struct AddHostSheet: View {
             ScrollViewReader { proxy in
             ScrollView {
                 VStack(spacing: 18) {
-                    hostSection
-                    monitoringSection
-                    credentialsSection
-                    testSection
-                    workingDirsSection
-                    tmuxConfSection
-                    scriptsSection
-                    agentModelsSection
-                        .id("agent-models")
-                    transportSection
+                    if showsModeBar {
+                        TallyChoiceBar(
+                            [("Bind", Mode.bind), ("Manual", Mode.manual)],
+                            selection: $mode
+                        )
+                    }
+                    switch resolvedMode {
+                    case .bind:
+                        BindPane()
+                    case .manual:
+                        manualForm
+                    }
                 }
                 .frame(maxWidth: 680)
                 .padding(18)
@@ -104,30 +120,87 @@ struct AddHostSheet: View {
             // fires on inert ground and never fights a control.
             .onTapGesture(perform: unfocusInputs)
             .task { await scrollForVerificationIfRequested(proxy) }
+            .task { openBindPaneForVerificationIfRequested() }
             .chassisSheetGround()
-            .navigationTitle(editing == nil ? "Add Host" : "Host Settings")
+            .navigationTitle(title)
             .navigationBarTitleDisplayMode(.inline)
             .toolbar {
-                ChassisSheetTitle(editing == nil ? "Add Host" : "Host Settings")
+                ChassisSheetTitle(title)
                 ToolbarItem(placement: .cancellationAction) {
-                    ChassisBarButton("Cancel") { clearSecretsAndDismiss() }
+                    // Nothing on the Bind pane is this modal's to save — the
+                    // machine and the controller finish that between them —
+                    // so the bar offers a way out, not a discard.
+                    ChassisBarButton(resolvedMode == .bind ? "Done" : "Cancel") {
+                        clearSecretsAndDismiss()
+                    }
                 }
-                ToolbarItem(placement: .confirmationAction) {
-                    ChassisBarButton("Save", action: save)
-                        .disabled(!isValid)
+                if resolvedMode == .manual {
+                    ToolbarItem(placement: .confirmationAction) {
+                        ChassisBarButton("Save", action: save)
+                            .disabled(!isValid)
+                    }
                 }
             }
             }
         }
-        // This long form is easy to drag while scrolling. Require the
-        // explicit Cancel or Save action so a swipe cannot discard edits;
-        // both paths also clear secret bindings before the sheet tears down.
-        .interactiveDismissDisabled()
+        // The manual form is long and easy to drag while scrolling. Require
+        // the explicit Cancel or Save action there so a swipe cannot discard
+        // edits; both paths also clear secret bindings before the sheet tears
+        // down. The Bind pane holds no unsaved edits, so a swipe is allowed —
+        // except mid-enrollment, where it would look like a cancel and isn't.
+        .interactiveDismissDisabled(
+            resolvedMode == .manual || bind.pending.contains(where: \.isBusy)
+        )
         .onAppear(perform: populate)
         .sheet(isPresented: $showingPaywall) { ProPaywallView() }
+        // A scan or paste that isn't a bind code at all. It belongs to this
+        // modal now: the deck has no bind surface to present it from, and an
+        // alert raised behind an open sheet would never be seen.
+        .alert(
+            "Bind Code Not Recognized",
+            isPresented: Binding(
+                get: { bind.alert != nil },
+                set: { if !$0 { bind.alert = nil } }
+            ),
+            presenting: bind.alert
+        ) { _ in
+            Button("OK", role: .cancel) {}
+        } message: { message in
+            Text(message)
+        }
+        // Binding creates a host, so it meets the same free-tier limit the
+        // manual form does — refused before anything is enrolled on the
+        // machine, never after.
+        .onChange(of: bind.needsProForHostLimit) { _, needsPro in
+            guard needsPro else { return }
+            bind.needsProForHostLimit = false
+            showingPaywall = true
+        }
         // Any edit that could change the outcome retires the shown result —
         // a stale PASSED next to a new address would vouch for the wrong host.
         .onChange(of: testFingerprint) { testState = .idle }
+    }
+
+    private var title: String { editing == nil ? "Add Host" : "Host Settings" }
+
+    /// Editing an existing host is form-only; there is no machine to bind.
+    private var showsModeBar: Bool { editing == nil }
+
+    private var resolvedMode: Mode { showsModeBar ? mode : .manual }
+
+    private var manualForm: some View {
+        VStack(spacing: 18) {
+            hostSection
+            monitoringSection
+            credentialsSection
+            testSection
+            workingDirsSection
+            tmuxConfSection
+            scriptsSection
+            agentModelsSection
+                .id("agent-models")
+            transportSection
+        }
     }
 
     // MARK: Form sections
@@ -785,6 +858,20 @@ struct AddHostSheet: View {
         ] == "models" else { return }
         try? await Task.sleep(for: .milliseconds(700))
         withAnimation(nil) { proxy.scrollTo("agent-models", anchor: .top) }
+        #endif
+    }
+
+    /// `MULTIPLEX_AUTO_ADD_HOST=bind|manual` picks which road this modal
+    /// opens on, so both panes can be captured headlessly (the simulator
+    /// cannot tap the choice bar). Adding always defaults to Bind.
+    private func openBindPaneForVerificationIfRequested() {
+        #if DEBUG
+        guard showsModeBar,
+              let request = ProcessInfo.processInfo.environment[
+                "MULTIPLEX_AUTO_ADD_HOST"
+              ]
+        else { return }
+        mode = request == "manual" ? .manual : .bind
         #endif
     }
